File name changes, Additional Folder, Added additional info files
Changed the names of each iteration of HW02 (adding each "Adjustment") to be version numbers.

Added an additional docs folder that contains images, excel files, and other documents that help me write the program.

The file ending with v.90 finally adds the ability to Rotate the icon. Changes to the order of asking the user for info on their icon. Made the terminal print more pleasing to the eyes.

~~~~~
Still need to add docstrings and comments. Probably v1.0 will have those.